# api/openapi/spec.yaml
openapi: "3.0.3"
info:
  title: Paily Connect Core
  version: "1.0.0"
  description: >
    Proxy node aggregation, filtering and distribution service.
    Fetcher and Checker authenticate with Bearer service-secret.
    Admin endpoints accept either a Bearer JWT (obtained via /api/v1/auth/login)
    or the shared service-secret Bearer token.
servers:
  - url: /
  # Default security: admin endpoints accept AdminJWT OR ServiceBearer (OR = separate list entries).
  # Public endpoints and ServiceBearer-only endpoints override this with their own security field.
security:
  - AdminJWT: []
  - ServiceBearer: []
components:
  securitySchemes:
    AdminJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ServiceBearer:
      type: http
      scheme: bearer
  schemas:
...
    PaginationInfo:
      type: object
      required: [page, limit, total, total_pages]
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, format: int64, example: 342 }
        total_pages: { type: integer, example: 18 }

    # ── Auth ─────────────────────────────────────────────────────────────────
...
    SourceSummary:
      type: object
      required: [id, type, identifier, info, status]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [subscribe, node] }
        identifier: { type: string }
        info: { type: string }
        status: { type: string, enum: [active, dead] }
        ignore_dead: { type: boolean }
        content: { type: string }
        latest_fetch_node_count:
          {
            type: integer,
            description: "Node count from the most recent successful fetch (subscribe sources only)",
          }
    SourceDetail:
      allOf:
        - $ref: "#/components/schemas/SourceSummary"
        - type: object
          properties:
            content: { type: string }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    SourceCreateRequest:
      type: object
      required: [type, content]
      properties:
        type: { type: string, enum: [subscribe, node] }
        identifier: { type: string }
        info: { type: string }
        content: { type: string }
    SourceUpdateRequest:
      type: object
      properties:
        identifier: { type: string }
        info: { type: string }
        content: { type: string }
...
    SourceListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          { type: array, items: { $ref: "#/components/schemas/SourceSummary" } }
        pagination: { $ref: "#/components/schemas/PaginationInfo" }
    SourceSearchRequest:
      type: object
      properties:
        expr: { type: string }
        page: { type: integer, default: 1 }
        limit: { type: integer, default: 20 }

    # ── Node ─────────────────────────────────────────────────────────────────
    NodeSummary:
      type: object
      required: [id, hash, server, protocol, region, alive, score]
      properties:
        id: { type: string, format: uuid }
        hash: { type: string }
        server: { type: string }
        protocol: { type: string }
        region: { type: string }
        alive: { type: boolean }
        score: { type: number, format: double }
        latest_latency_ms: { type: integer }
        latest_avg_latency_ms: { type: integer }
        latest_jitter_ms: { type: integer }
        sources:
          type: array
          items: { $ref: "#/components/schemas/SourceSummary" }
...
    NodeListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          { type: array, items: { $ref: "#/components/schemas/NodeSummary" } }
        pagination: { $ref: "#/components/schemas/PaginationInfo" }
    NodeSearchRequest:
      type: object
      properties:
        expr: { type: string }
        page: { type: integer, default: 1 }
        limit: { type: integer, default: 20 }
...
    FetchSourcesResponse:
      type: object
      required: [sources]
      properties:
        sources:
          type: array
          items: { $ref: "#/components/schemas/SourceDetail" }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
paths:
...
  /api/v1/fetch/sources:
    get:
      operationId: FetchGetSources
      tags: [fetch]
      summary: Get all active sources for fetcher
      security: [{ ServiceBearer: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FetchSourcesResponse" }
        "401": { description: Unauthorized }
...
  /api/v1/sources:
    post:
      operationId: SourceCreate
      tags: [sources]
      summary: Catcher submits a new source (deduped by type+content)
      security: [{ ServiceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceCreateRequest" }
      responses:
        "200": { description: Already exists (no-op) }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceDetail" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { description: Unauthorized }
    get:
      operationId: SourceList
      tags: [sources]
      summary: List sources with optional filters
      security: [{ AdminJWT: [] }, { ServiceBearer: [] }]
      parameters:
        - {
            name: type,
            in: query,
            schema: { type: string, enum: [subscribe, node] },
          }
        - {
            name: status,
            in: query,
            schema: { type: string, enum: [active, dead] },
          }
        - { name: identifier, in: query, schema: { type: string } }
        - { name: info, in: query, schema: { type: string } }
        - {
            name: sort_by,
            in: query,
            schema: { type: string, enum: [node_count] },
            description: "Sort field (default: created_at desc)",
          }
        - {
            name: sort_dir,
            in: query,
            schema: { type: string, enum: [asc, desc], default: desc },
            description: "Sort direction",
          }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceListResponse" }
        "401": { description: Unauthorized }
  /api/v1/sources/search:
    post:
      operationId: SourceSearch
      tags: [sources]
      summary: Search sources using an expr expression
      security: [{ AdminJWT: [] }, { ServiceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceSearchRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceListResponse" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { description: Unauthorized }
  /api/v1/sources/{id}:
    parameters:
      - {
          name: id,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      operationId: SourceGet
      tags: [sources]
      summary: Get source detail
      security: [{ AdminJWT: [] }, { ServiceBearer: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceDetail" }
        "401": { description: Unauthorized }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    put:
      operationId: SourceUpdate
      tags: [sources]
      summary: Update source identifier/info
      security: [{ AdminJWT: [] }, { ServiceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceDetail" }
        "401": { description: Unauthorized }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      operationId: SourceDelete
      tags: [sources]
      summary: Delete a source and cascade-remove orphaned nodes
      security: [{ AdminJWT: [] }, { ServiceBearer: [] }]
      responses:
        "204": { description: Deleted }
        "401": { description: Unauthorized }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/sources/{id}/nodes:
    get:
      operationId: SourceGetNodes
      tags: [sources]
      summary: List nodes contributed by this source (traceability)
      security: [{ AdminJWT: [] }, { ServiceBearer: [] }]
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeListResponse" }
        "401": { description: Unauthorized }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/nodes/search:
    post:
      operationId: NodeSearch
      tags: [nodes]
      summary: Search nodes using an expr expression
      security: [{ AdminJWT: [] }, { ServiceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NodeSearchRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeListResponse" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { description: Unauthorized }
...